+++ processor/src/main/java/react4j/processor/Constants.java
@@ -7,6 +7,7 @@ final class Constants
   static final String PROCEDURE_CLASSNAME = "react4j.core.Procedure";
   static final String EVENT_HANDLER_ANNOTATION_CLASSNAME = "react4j.annotations.EventHandler";
   static final String REACT_COMPONENT_ANNOTATION_CLASSNAME = "react4j.annotations.ReactComponent";
+  static final String INJECT_ANNOTATION_CLASSNAME = "javax.inject.Inject";
 
   private Constants()
   {