+++ processor/src/test/resources/expected/com/example/inject/ArezReactComponent_.java
@@ -1,0 +1,95 @@
+package com.example.inject;
+
+import javax.annotation.Generated;
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import jsinterop.annotations.JsType;
+import jsinterop.base.JsPropertyMap;
+import org.realityforge.arez.annotations.ArezComponent;
+import react4j.core.BaseContext;
+import react4j.core.BaseProps;
+import react4j.core.BaseState;
+import react4j.core.ComponentConstructorFunction;
+import react4j.core.NativeAdapterComponent;
+import react4j.core.React;
+import react4j.core.ReactConfig;
+import react4j.core.ReactNode;
+
+@ArezComponent(
+    type = "ArezReactComponent",
+    inject = true
+)
+@Generated("react4j.processor.ReactProcessor")
+class ArezReactComponent_ extends ArezReactComponent {
+  private static final ComponentConstructorFunction<BaseProps, BaseContext> TYPE = getConstructorFunction();
+
+  @Nonnull
+  static ReactNode _create() {
+    return React.createElement( TYPE );
+  }
+
+  @Nonnull
+  static ReactNode _create(@Nullable final BaseProps props) {
+    return React.createElement( TYPE, props );
+  }
+
+  @Nonnull
+  static ReactNode _create(@Nullable final BaseProps props, @Nullable final ReactNode child) {
+    return React.createElement( TYPE, props, child );
+  }
+
+  @Nonnull
+  private static ComponentConstructorFunction<BaseProps, BaseContext> getConstructorFunction() {
+    final ComponentConstructorFunction<BaseProps, BaseContext> componentConstructor = NativeReactComponent::new;
+    if ( ReactConfig.enableComponentNames() ) {
+      JsPropertyMap.of( componentConstructor ).set( "displayName", "ArezReactComponent" );
+    }
+    return componentConstructor;
+  }
+
+  @JsType(
+      isNative = true
+  )
+  interface Lifecycle {
+    void componentDidMount();
+
+    void componentDidUpdate(@Nonnull BaseProps arg0, @Nonnull BaseState arg1);
+
+    void componentWillUnmount();
+
+    boolean shouldComponentUpdate(@Nonnull BaseProps arg0, @Nonnull BaseState arg1,
+        @Nonnull BaseContext arg2);
+  }
+
+  static final class NativeReactComponent extends NativeAdapterComponent<BaseProps, BaseState, BaseContext, ArezReactComponent> implements Lifecycle {
+    NativeReactComponent(@Nonnull final BaseProps props, @Nonnull final BaseContext context) {
+      super( props, context );
+    }
+
+    @Override
+    protected ArezReactComponent createComponent() {
+      return new Arez_ArezReactComponent_();
+    }
+
+    @Override
+    public void componentDidMount() {
+      performComponentDidMount();
+    }
+
+    @Override
+    public void componentDidUpdate(@Nonnull final BaseProps arg0, @Nonnull final BaseState arg1) {
+      performComponentDidUpdate(arg0,arg1);
+    }
+
+    @Override
+    public void componentWillUnmount() {
+      performComponentWillUnmount();
+    }
+
+    @Override
+    public boolean shouldComponentUpdate(@Nonnull final BaseProps arg0,
+        @Nonnull final BaseState arg1, @Nonnull final BaseContext arg2) {
+      return performShouldComponentUpdate(arg0,arg1,arg2);
+    }
+  }
+}